[bugfix] Move instance creation methods to Assembly
