+++ src/abaqus/Assembly/PartInstance.py
@@ -189,194 +189,6 @@ class PartInstance:
         self.elementEdges = part.elementEdges
 
     @abaqus_method_doc
-    def InstanceFromBooleanCut(
-        self,
-        name: str,
-        instanceToBeCut: str,
-        cuttingInstances: Sequence["PartInstance"],
-        originalInstances: Literal[C.SUPPRESS, C.DELETE] = SUPPRESS,
-    ):
-        """This method creates a PartInstance in the instances repository after subtracting or cutting the
-        geometries of a group of part instances from that of a base part instance.
-
-        .. note::
-            This function can be accessed by::
-
-                mdb.models[name].rootAssembly.Instance
-
-        Parameters
-        ----------
-        name
-            A String specifying the repository key. The name must be a valid Abaqus object name.
-        instanceToBeCut
-            A PartInstance specifying the base instance from which to cut other instances.
-        cuttingInstances
-            A sequence of PartInstance objects specifying the instances with which to cut the base
-            instance.
-        originalInstances
-            A SymbolicConstant specifying whether the original instances should be suppressed or
-            deleted after the merge operation. Possible values are SUPPRESS or DELETE. The default
-            value is SUPPRESS.
-
-        Returns
-        -------
-        PartInstance
-            A PartInstance object.
-        """
-        ...
-
-    @abaqus_method_doc
-    def InstanceFromBooleanMerge(
-        self,
-        name: str,
-        instances: Sequence["PartInstance"],
-        keepIntersections: Boolean = False,
-        originalInstances: Literal[C.SUPPRESS, C.DELETE] = SUPPRESS,
-        domain: Literal[C.MESH, C.BOTH, C.GEOMETRY] = GEOMETRY,
-        mergeNodes: Literal[C.MESH, C.ALL, C.BOUNDARY_ONLY, C.NONE] = BOUNDARY_ONLY,
-        nodeMergingTolerance: Optional[float] = None,
-        removeDuplicateElements: Boolean = True,
-    ):
-        """This method creates a PartInstance in the instances repository after merging two or more part
-        instances.
-
-        .. note::
-            This function can be accessed by::
-
-                mdb.models[name].rootAssembly.Instance
-
-        Parameters
-        ----------
-        name
-            A String specifying the repository key. The name must be a valid Abaqus object name.
-        instances
-            A sequence of PartInstance objects specifying the part instances to merge.
-        keepIntersections
-            A Boolean specifying whether the boundary intersections of Abaqus native part instances
-            should be retained after the merge operation. The default value is False.
-        originalInstances
-            A SymbolicConstant specifying whether the original instances should be suppressed or
-            deleted after the merge operation. Possible values are SUPPRESS or DELETE. The default
-            value is SUPPRESS.
-        domain
-            A SymbolicConstant specifying whether geometry or mesh of the specified part instances
-            is to be merged. Possible values are GEOMETRY, MESH or BOTH. The default value is
-            GEOMETRY.
-        mergeNodes
-            A SymbolicConstant specifying which nodes of the specified part instances should be
-            considered for merging. This argument is only applicable if **domain** is MESH. Possible
-            values are BOUNDARY_ONLY, ALL, or NONE. The default value is BOUNDARY_ONLY.
-        nodeMergingTolerance
-            A Float specifying the maximum distance between nodes of the specified part instances
-            that will be merged and replaced with a single node in the new part. The location of the
-            new node is the average position of the deleted nodes. This argument is only applicable
-            if **domain** is MESH. The default value is 10⁻⁶.
-        removeDuplicateElements
-            A Boolean specifying whether elements with the same connectivity in the new part will be
-            merged into a single element. This argument is only applicable if **domain** is MESH. The
-            default value is True.
-
-        Returns
-        -------
-        PartInstance
-            A PartInstance object.
-        """
-        ...
-
-    @abaqus_method_doc
-    def LinearInstancePattern(
-        self,
-        instanceList: tuple,
-        number1: int,
-        spacing1: float,
-        number2: int,
-        spacing2: float,
-        direction1: tuple = (),
-        direction2: tuple = (),
-    ) -> Sequence["PartInstance"]:
-        """This method creates multiple PartInstance objects in a linear pattern and puts them into the
-        instances repository.
-
-        .. note::
-            This function can be accessed by::
-
-                mdb.models[name].rootAssembly.Instance
-
-        Parameters
-        ----------
-        instanceList
-            A sequence of Strings specifying the names of instances to pattern.
-        number1
-            An Int specifying the total number of instances, including the original instances, that
-            appear along the first direction in the pattern.
-        spacing1
-            A Float specifying the spacing between instances along the first direction in the
-            pattern.
-        number2
-            An Int specifying the total number of instances, including the original instances, that
-            appear along the second direction in the pattern.
-        spacing2
-            A Float specifying the spacing between instances along the second direction in the
-            pattern.
-        direction1
-            A sequence of three Floats specifying a vector along the first direction. The default
-            value is (1.0, 0.0, 0.0).
-        direction2
-            A sequence of three Floats specifying a vector along the second direction. The default
-            value is (0.0, 1.0, 0.0).
-
-        Returns
-        -------
-        Sequence[PartInstance]
-            A sequence of PartInstance objects.
-        """
-        ...
-
-    @abaqus_method_doc
-    def RadialInstancePattern(
-        self,
-        instanceList: tuple,
-        number: int,
-        totalAngle: float,
-        point: Sequence[float] = (),
-        axis: Sequence[float] = (),
-    ):
-        """This method creates multiple PartInstance objects in a radial pattern and puts them into the
-        instances repository.
-
-        .. note::
-            This function can be accessed by::
-
-                mdb.models[name].rootAssembly.Instance
-
-        Parameters
-        ----------
-        instanceList
-            A sequence of Strings specifying the names of instances to pattern.
-        number
-            An Int specifying the total number of instances, including the original instances, that
-            appear in the radial pattern.
-        totalAngle
-            A Float specifying the total angle in degrees between the first and last instance in the
-            pattern. A positive angle corresponds to a counter-clockwise direction. The values 360°
-            and -360° represent a special case where the pattern makes a full circle. In this case,
-            because the copy would overlay the original, the copy is not placed at the last
-            position. Possible values are -360.0 ≤ **totalAngle** ≤ 360.0.
-        point
-            A sequence of three Floats specifying the center of the radial pattern. The default
-            value is (0.0, 0.0, 0.0).
-        axis
-            A sequence of three Floats specifying the central axis of the radial pattern. The
-            default value is (0.0, 0.0, 1.0).
-
-        Returns
-        -------
-        Sequence[PartInstance]
-            A sequence of PartInstance objects.
-        """
-        ...
-
-    @abaqus_method_doc
     def checkGeometry(self, detailed: Boolean = OFF, level: Optional[int] = None):
         """This method checks the validity of the geometry of the part instance and prints a count of all
         topological entities on the part instance (faces, edges, vertices, etc.).
